# swagger.yaml
openapi: 3.0.0
info:
  title: WhSmith
  contact: {}
  version: '1.0'
servers:
- url: https://avnology.org
  variables: {}
- url: http://localhost:1812
  variables: {}
paths:
...
  /auth/verifySignup:
    post:
      tags:
      - Authentication
      summary: Verify For Signup
      operationId: VerifyForSignup
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/VerifyForSignupRequest'
              - example:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWRGb3JTaWdudXAiOjEsInVzZXJJZCI6IjY1MzdkOGZhNTYzYzI4NjliYzQ2NTRmNCIsImlhdCI6MTY5ODM3MTA4NiwiZXhwIjoxNjk4Mzc0Njg2fQ.RqglWDpaJ6C9uP8URnjCBhrvo8RzvBR6GZh6PZ-p89g
            example:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWRGb3JTaWdudXAiOjEsInVzZXJJZCI6IjY1MzdkOGZhNTYzYzI4NjliYzQ2NTRmNCIsImlhdCI6MTY5ODM3MTA4NiwiZXhwIjoxNjk4Mzc0Njg2fQ.RqglWDpaJ6C9uP8URnjCBhrvo8RzvBR6GZh6PZ-p89g
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /auth/forgetPassword:
    post:
      tags:
      - Authentication
      summary: Forget Password
      operationId: ForgetPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/ForgetPasswordRequest'
              - example:
                  email: mh3908334@gmail.com
            example:
              email: mh3908334@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /auth/verifyPassword:
    post:
      tags:
      - Authentication
      summary: Verify Password
      operationId: VerifyPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/VerifyPasswordRequest'
              - example:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NTE5N2NkMTczYjBhNTk1YWQ5ZmFiYWIiLCJpYXQiOjE2OTYxNzA3MzEsImV4cCI6MTY5ODc2MjczMX0.T37VjcXf34Md_TTo3_C7Yc2eTsZk-eKR6KfDHqjeLn4
            example:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NTE5N2NkMTczYjBhNTk1YWQ5ZmFiYWIiLCJpYXQiOjE2OTYxNzA3MzEsImV4cCI6MTY5ODc2MjczMX0.T37VjcXf34Md_TTo3_C7Yc2eTsZk-eKR6KfDHqjeLn4
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /auth/resetPassword:
    patch:
      tags:
      - Authentication
      summary: Reset Password
      operationId: ResetPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/ResetPasswordRequest'
              - example:
                  email: mh3908334@gmail.com
                  newPassword: '123456'
            example:
              email: mh3908334@gmail.com
              newPassword: '123456'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /auth/changeEmail:
    post:
      tags:
      - Authentication
      summary: Change Email
      operationId: ChangeEmail
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/ChangeEmailRequest'
              - example:
                  email: mh3908334@gmail.com
                  newEmail: ads@gmail.com
            example:
              email: mh3908334@gmail.com
              newEmail: ads@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    VerifyForSignupRequest:
      title: VerifyForSignupRequest
      required:
      - token
      type: object
      properties:
        token:
          type: string
      example:
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWRGb3JTaWdudXAiOjEsInVzZXJJZCI6IjY1MzdkOGZhNTYzYzI4NjliYzQ2NTRmNCIsImlhdCI6MTY5ODM3MTA4NiwiZXhwIjoxNjk4Mzc0Njg2fQ.RqglWDpaJ6C9uP8URnjCBhrvo8RzvBR6GZh6PZ-p89g
    ForgetPasswordRequest:
      title: ForgetPasswordRequest
      required:
      - email
      type: object
      properties:
        email:
          type: string
      example:
        email: mh3908334@gmail.com
    VerifyPasswordRequest:
      title: VerifyPasswordRequest
      required:
      - token
      type: object
      properties:
        token:
          type: string
      example:
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NTE5N2NkMTczYjBhNTk1YWQ5ZmFiYWIiLCJpYXQiOjE2OTYxNzA3MzEsImV4cCI6MTY5ODc2MjczMX0.T37VjcXf34Md_TTo3_C7Yc2eTsZk-eKR6KfDHqjeLn4
    ResetPasswordRequest:
      title: ResetPasswordRequest
      required:
      - email
      - newPassword
      type: object
      properties:
        email:
          type: string
        newPassword:
          type: string
      example:
        email: mh3908334@gmail.com
        newPassword: '123456'
    ChangeEmailRequest:
      title: ChangeEmailRequest
      required:
      - email
      - newEmail
      type: object
      properties:
        email:
          type: string
        newEmail:
          type: string
      example:
        email: mh3908334@gmail.com
        newEmail: ads@gmail.com
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...